fix(test): correct phantom + window boundary tests; split phantom tests file (#82)
Three issues from previous run 25425008033:

1. `phantom_cluster_gap_after_at_threshold_drops` was inverted. Code is
   `if gap_after < 3000 → skip`, so gap_after EXACTLY 3000 fails the
   strict-less guard and DROPS. Test was asserting no-drop. Fixed: at
   3000 → drops (boundary inclusive); at 2999 → does not drop.

2. `best_window_match_finds_pair_when_window_grows_past_start_plus_one`
   used asr [a@0-100, b@200-300] giving span 300 ms — below
   MIN_LINE_DURATION_MS (500). Caught the matched pair but bounced off
   the span guard, returning None and failing the assertion. Fixed:
   moved b to 700-1000 → span 1000 ms, well over the floor.

3. description_merge_tests.rs grew to 1290 lines after the phantom
   batch — over the 1000-line cap. Split phantom-related tests (~290
   lines, 16 cases including the boundary additions) into a new
   sibling file `description_merge_phantom_tests.rs` and registered
   it as a second `#[cfg(test)] mod phantom_tests` from
   description_merge.rs alongside the existing `mod tests`.

description_merge.rs sat at 1001 lines after adding the new mod
declaration; trimmed a constant doc-comment to 1000 (cap is `> 1000`,
so 1000 passes).
